% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildAdjacencyMatrix.R
\name{buildAdjacencyMatrix}
\alias{buildAdjacencyMatrix}
\title{Build an Adjacency matrix of users to web domains}
\usage{
buildAdjacencyMatrix(data)
}
\arguments{
\item{data}{A dataframe consisting of at least 4 columns: Authority_URI, HTTPUSERAGENT, LATITUDE_SRC, LONGITUDE_SRC}
}
\value{
adjacencymatrix An ``nxn'' sparse matrix from the Matrix Package of all users connected to web domains
}
\description{
This function takes as input a column that has user information and location sourcing as well as the web sites visited. It then returns a sparse binary adjacency matrix where users are rows, columns are web domains, and entries are indicators of if the user visited that web domain
}
\details{
Build the Adjacency matrix
}
\examples{


}
