% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterCopurchaseMatrix.R
\name{filterCopurchaseMatrix}
\alias{filterCopurchaseMatrix}
\title{Get Top n Values of a Matrix}
\usage{
filterCopurchaseMatrix(copurchase.matrix, n = 3, filterlevel = 0.05)
}
\arguments{
\item{copurchase.matrix}{A sparse Matrix connecting web domains to other web domains by common users}

\item{n}{The top n heaviest weighted edges to average}

\item{filterlevel}{The percentage level chosen to reduce the top n heaviest edges}
}
\value{
a filtered sparse Copurchase matrix
}
\description{
This function imports a sparse adjacency matrix. It then applies the Top-n Heaviest Edges Filtering method needed to remove spurious connections between web domains
}
