% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCommunities.R
\name{getCommunities}
\alias{getCommunities}
\title{get Communities}
\usage{
getCommunities(copurchase.matrix)
}
\arguments{
\item{copurchase.matrix}{a sparse nxn matrix connecting web domains by common users which will be turned into a graph for community detection}
}
\value{
community.walktrap An igraph object containing community definitions for each web domain
}
\description{
This function imports a copurchase matrix, usually after filtering, and builds a graph and conducts community detection on the graph using the walktrap method. It then returns that walktrap object
}
