% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsetoMatrix.R
\name{getTopNvalues}
\alias{getTopNvalues}
\title{get Top N Values}
\usage{
getTopNvalues(x, n = 3)
}
\arguments{
\item{x}{A sparse matrix object}

\item{n}{how many highest number of entries to keep}
}
\value{
A vector of the top n highest entries
}
\description{
get Top N Values
}
