% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadWebClicks.R
\name{loadWebClicks}
\alias{loadWebClicks}
\title{Load in Web Click Data}
\usage{
loadWebClicks(data.location)
}
\arguments{
\item{data.location}{String File location name}
}
\value{
A DataFrame consisting of the important columns from the web click data
}
