% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_my_app.R
\name{run_my_app}
\alias{run_my_app}
\title{Run an app in my package}
\usage{
run_my_app(app_name, ...)
}
\arguments{
\item{app_name}{character string for the package you want to run}

\item{...}{Additional options pased to shinyAppDir}
}
\value{
A printed shiny app
}
\description{
Use `run_my_app` to start and use apps located within in the package
}
