% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparsetoMatrix.R
\name{sparseToMatrix}
\alias{sparseToMatrix}
\title{Convert a sparse matrix to a vector representation}
\usage{
sparseToMatrix(x)
}
\arguments{
\item{x}{a sparse Matrix}
}
\value{
a vector from 1:n of weights
}
\description{
This function takes in a sparse matrix and returns a vector of all entries. This allows the top three heaviest edges to work
}
