% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWAT.R
\docType{package}
\name{SWAT-package}
\alias{SWAT}
\alias{SWAT-package}
\title{SWAT: Shortened Web Link Analysis Tool}
\description{
The SWAT Package provides a convenient wrapper for the 
SWAT Shiny App. To load the app, call `SWAT::run_my_app("SWAT")`
}
\author{
\strong{Maintainer}: James Gallagher \email{james.c.gallagher@gmail.com}

}
