% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webclickdataexample.R
\docType{data}
\name{webclickexample}
\alias{webclickexample}
\title{webclick data example}
\format{A tibble with 35 columns}
\usage{
webclickexample
}
\description{
This is an example of web click data
}
\keyword{datasets}
