% Generated by roxygen2: do not edit by hand
2

% Please edit documentation in R/cleanData.R
3

\name{cleanData}
4

\alias{cleanData}
5

\title{Clean web click data}
6

\usage{
  7
  
  cleanData(data, matchvector)
  8
  
}
9

\arguments{
  10
  
  \item{data}{A dataframe pulled from the webclick data}
  11
  
  12
  
  \item{matchvector}{a vector of strings used to identify topics of link clicks}
  13
  
}
14

\value{
  15
  
  datanoads a dataframe with ads removed and prepared for analysis
  16
  
}
17

\description{
  18
  
  This function takes a link click dataset and removes extraneous clicks directed to ads, ip addresses, etc.
  19
  
}